Remove bpf_loader.rs
