--- src/bpf_loader.rs
@@ -1,24 +1,0 @@
-//! BPF loader
-use solana_sdk::account::Account;
-use solana_sdk::native_loader;
-use solana_sdk::pubkey::Pubkey;
-
-const BPF_LOADER_NAME: &str = "solana_bpf_loader";
-const BPF_LOADER_PROGRAM_ID: [u8; 32] = [
-    128, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0,
-];
-
-pub fn id() -> Pubkey {
-    Pubkey::new(&BPF_LOADER_PROGRAM_ID)
-}
-
-pub fn account() -> Account {
-    Account {
-        tokens: 1,
-        owner: id(),
-        userdata: BPF_LOADER_NAME.as_bytes().to_vec(),
-        executable: true,
-        loader: native_loader::id(),
-    }
-}
